refactor reservation flow

+++ app/src/main/java/com/tr/helpark/helparkcapstoneproject/features/home/data/dto/response/GetAllParksResponseDto.kt
@@ -44,7 +44,7 @@ data class GetAllParksResponseDtoItem(
     @SerializedName("resTime")
     val resTime: Int?,
     @SerializedName("hire")
-    val hire: Int?
+    val hire: Float?
 )
 
 data class ParkDetail(
